requirements.txt
pyyaml
jinja2
requests
python-slugify
